Removed isLazy property in ParseUser
We always have the invariant:
lazy == true <==> objectId == null && authData.get("anonymous") != null

So we can remove the additional property.

For testing: to get the "lazy" user behavior we need to make sure that the user has no objectId and we need to set "anonymous" authData (see setLazy() in ParseACLTest.java).