Include dependency & references shield in README
Add a dependency shield that determines if the dependencies in the repository are up to date. If not, the shield will indicate out of date and this will notify contributors that the libraries need to be updated.

In addition, a references shield is included which shows the number of projects that reference this project.

These two badges are included in the Android-SDK-iOS-OSX project and it is nice to stay consistent between the various projects of Parse. Badges are obtained via VersionEye.